+++ Backend/modelanalysisv2.h
@@ -14,28 +14,28 @@ class ModelAnalysisV2 : public QObject
 {
     Q_OBJECT
 
-    Q_PROPERTY(float mae READ mae NOTIFY maeChanged)
-    Q_PROPERTY(float mape READ mape NOTIFY mapeChanged)
-    Q_PROPERTY(float mse READ mse NOTIFY mseChanged)
-    Q_PROPERTY(float rmse READ rmse NOTIFY rmseChanged)
+    Q_PROPERTY(double mae READ mae NOTIFY maeChanged)
+    Q_PROPERTY(double mape READ mape NOTIFY mapeChanged)
+    Q_PROPERTY(double mse READ mse NOTIFY mseChanged)
+    Q_PROPERTY(double rmse READ rmse NOTIFY rmseChanged)
 
-    Q_PROPERTY(float mae2 READ mae2 NOTIFY mae2Changed)
-    Q_PROPERTY(float mape2 READ mape2 NOTIFY mape2Changed)
-    Q_PROPERTY(float mse2 READ mse2 NOTIFY mse2Changed)
-    Q_PROPERTY(float rmse2 READ rmse2 NOTIFY rmse2Changed)
+    Q_PROPERTY(double mae2 READ mae2 NOTIFY mae2Changed)
+    Q_PROPERTY(double mape2 READ mape2 NOTIFY mape2Changed)
+    Q_PROPERTY(double mse2 READ mse2 NOTIFY mse2Changed)
+    Q_PROPERTY(double rmse2 READ rmse2 NOTIFY rmse2Changed)
 
 public:
     explicit ModelAnalysisV2(QObject *parent = nullptr, const std::vector<DataRow2>& data = {});
 
-    float mae() const;
-    float mape() const;
-    float mse() const;
-    float rmse() const;
+    double mae() const;
+    double mape() const;
+    double mse() const;
+    double rmse() const;
 
-    float mae2() const;
-    float mape2() const;
-    float mse2() const;
-    float rmse2() const;
+    double mae2() const;
+    double mape2() const;
+    double mse2() const;
+    double rmse2() const;
 
     Q_INVOKABLE void calculate2();
 
@@ -59,15 +59,15 @@ private:
 private:
     std::vector<DataRow2> m_FileData2;
 
-    float m_MAE;
-    float m_MAPE;
-    float m_MSE;
-    float m_RMSE;
+    double m_MAE;
+    double m_MAPE;
+    double m_MSE;
+    double m_RMSE;
 
-    float m_MAE2;
-    float m_MAPE2;
-    float m_MSE2;
-    float m_RMSE2;
+    double m_MAE2;
+    double m_MAPE2;
+    double m_MSE2;
+    double m_RMSE2;
 
 };
